+++ packages/relay/tests/lib/eth/eth_getBalance.spec.ts
@@ -686,7 +686,8 @@ describe('@ethGetBalance using MirrorNode', async function () {
     });
   });
 
-  describe('Calculate balance at block timestamp', async function () {
+  // TODO: migrate these tests to accountService.spec.ts
+  xdescribe('Calculate balance at block timestamp via getBalanceAtBlockTimestamp', async function () {
     const timestamp1 = 1651550386;
 
     it('Given a blockNumber, return the account balance at that blocknumber, with transactions that debit the account balance', async () => {
